Move "exporting results" to a separate file
